+++ .pylintrc
@@ -225,7 +225,7 @@ indent-after-paren=4
 
 # String used as indentation unit. This is usually "    " (4 spaces) or "\t" (1
 # tab).
-indent-string='  '
+indent-string='    '
 
 # Maximum number of characters on a single line.
 max-line-length=100
