+++ Pipfile
@@ -17,6 +17,4 @@ Flask-SQLAlchemy = "*"
 passlib = "*"
 bcrypt = "*"
 Flask-Migrate = "*"
-awesome-slugify = "*"
-psycopg2 = "*"
 requests = "*"
