--- Procfile
@@ -1,1 +1,0 @@
-web: python src/app.py
