+++ src/db_helper.py
@@ -1,7 +1,3 @@
-"""
-Database Interface
-"""
-
 from models import db
 
 IS_DESTROYED = 'is_destroyed'
